Multi-loci heuristic

Closes #3381, #5114, #4962, #2259, #3740, #4937, #4942, #4887, #4859, #4818, #4844, #4773, #4742, #4775, #4425, #4722, #4701, #4725, #4660, #4727, and #4338

See merge request vidjil/vidjil!1527

+++ algo/core/proba.h
@@ -30,7 +30,7 @@ private:
    * Commpute the previous iteration for the probability (and updates the values Cnk, proba_with,
    * proba_without.
    */
-  inline double probabilityPreviousIteration(int iteration, int length, double &Cnk, double &proba_with, double &proba_without, float index_load);
+  inline double probabilityPreviousIteration(int iteration, int length, long double &Cnk, double &proba_with, double &proba_without, float index_load);
 public:
